+++ src/OpenLoco/src/Graphics/SoftwareDrawingEngine.cpp
@@ -414,4 +414,19 @@ namespace OpenLoco::Gfx
         return _screenInfo;
     }
 
+    bool SoftwareDrawingEngine::setVSync(bool state)
+    {
+        if (_vsync == state)
+        {
+            return true;
+        }
+
+        if (SDL_RenderSetVSync(_renderer, state ? 1 : 0) == 0)
+        {
+            _vsync = state;
+            return true;
+        }
+
+        return false;
+    }
 }